refactored CommitteeSession

+++ app/Http/Controllers/CommitteeSessionController.php
@@ -9,7 +9,8 @@ use App\Http\Requests\CommitteeSession\UpdateRequest;
 use App\Http\Resources\CommitteeSessionResource;
 use App\Http\Resources\CommitteeSessionResourceCollection;
 use App\Repositories\Contracts\CommitteeSessionRepository;
-use Illuminate\Http\Request;
+use Illuminate\Auth\Access\AuthorizationException;
+use Illuminate\Http\Response;
 
 class CommitteeSessionController extends Controller
 {
@@ -32,9 +33,12 @@ class CommitteeSessionController extends Controller
      *
      * @param IndexRequest $request
      * @return CommitteeSessionResourceCollection
+     * @throws AuthorizationException
      */
     public function index(IndexRequest $request)
     {
+        $this->authorize('list', [CommitteeSession::class, $request->get('propertyId')]);
+
         $committeeSessions = $this->committeeSessionRepository->findBy($request->all());
 
         return new CommitteeSessionResourceCollection($committeeSessions);
@@ -43,11 +47,14 @@ class CommitteeSessionController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param StoreRequest $request
      * @return CommitteeSessionResource
+     * @throws AuthorizationException
      */
     public function store(StoreRequest $request)
     {
+        $this->authorize('store', [CommitteeSession::class, $request->get('propertyId')]);
+
         $committeeSession = $this->committeeSessionRepository->save($request->all());
 
         return new CommitteeSessionResource($committeeSession);
@@ -58,9 +65,12 @@ class CommitteeSessionController extends Controller
      *
      * @param CommitteeSession $committeeSession
      * @return CommitteeSessionResource
+     * @throws AuthorizationException
      */
     public function show(CommitteeSession $committeeSession)
     {
+        $this->authorize('show', $committeeSession);
+
         return new CommitteeSessionResource($committeeSession);
     }
 
@@ -70,9 +80,12 @@ class CommitteeSessionController extends Controller
      * @param UpdateRequest $request
      * @param CommitteeSession $committeeSession
      * @return CommitteeSessionResource
+     * @throws AuthorizationException
      */
     public function update(UpdateRequest $request, CommitteeSession $committeeSession)
     {
+        $this->authorize('update', $committeeSession);
+
         $committeeSession = $this->committeeSessionRepository->update($committeeSession, $request->all());
 
         return new CommitteeSessionResource($committeeSession);
@@ -82,10 +95,13 @@ class CommitteeSessionController extends Controller
      * Remove the specified resource from storage.
      *
      * @param CommitteeSession $committeeSession
-     * @return \Illuminate\Http\Response
+     * @return Response
+     * @throws AuthorizationException
      */
     public function destroy(CommitteeSession $committeeSession)
     {
+        $this->authorize('destroy', $committeeSession);
+
         $this->committeeSessionRepository->delete($committeeSession);
 
         return response()->json(null, 204);
